refactor: UserFixture를 User 하위로 이동
- 코드 컨벤션에 따른 코드 수정도 포함
